requirements.txt
streamlit
langchain
langgraph
langchain-ollama
langchain-groq
langchain-core
langchain-community
faiss-cpu
pypdf2
docx
PyMuPDF 
Pillow